lint
SCCS-vsn: sys/kern/kern_clock.c 4.42
SCCS-vsn: sys/kern/kern_proc.c 4.43
SCCS-vsn: sys/sys/kernel.h 4.7
SCCS-vsn: sys/kern/kern_sig.c 5.10
SCCS-vsn: sys/kern/kern_synch.c 4.24
SCCS-vsn: sys/kern/subr_rmap.c 4.7
SCCS-vsn: sys/kern/subr_rmap.c.sav 4.7
SCCS-vsn: sys/kern/sys_generic.c 5.20
SCCS-vsn: sys/vax/vax/conf.c 4.59
SCCS-vsn: sys/kern/uipc_syscalls.c 4.35
SCCS-vsn: sys/vax/vax/flp.c 4.11
SCCS-vsn: sys/vax/vax/mem.c 4.9
SCCS-vsn: sys/ufs/ffs/ffs_alloc.c 2.18
SCCS-vsn: sys/ufs/lfs/lfs_alloc.c 2.18
SCCS-vsn: sys/vax/vax/trap.c 4.21
SCCS-vsn: sys/ufs/ffs/ffs_subr.c 4.2
SCCS-vsn: sys/kern/uipc_mbuf.c 1.39
SCCS-vsn: sys/kern/uipc_socket.c 4.59
SCCS-vsn: sys/vm/vm_swap.c 4.10
SCCS-vsn: sys/kern/kern_physio.c 4.34
SCCS-vsn: sys/vax/if/if_acc.c 4.24
SCCS-vsn: sys/vax/if/if_ec.c 4.25
SCCS-vsn: sys/deprecated/netdecnet/nsp_output.c 1.4
SCCS-vsn: sys/deprecated/netdecnet/nsp_usrreq.c 1.4
SCCS-vsn: sys/netinet/in.c 4.9
SCCS-vsn: sys/netinet/ip_input.c 1.55
SCCS-vsn: sys/netinet/tcp_subr.c 4.33
SCCS-vsn: sys/netinet/udp_usrreq.c 4.38
SCCS-vsn: sys/vax/uba/dmf.c 4.13
SCCS-vsn: sys/vax/uba/dn.c 4.13
SCCS-vsn: sys/vax/uba/ikreg.h 4.3
SCCS-vsn: sys/vax/uba/uba.c 4.54
SCCS-vsn: sys/vax/uba/uda.c 4.14
SCCS-vsn: sys/deprecated/netimp/if_imp.c 4.42
SCCS-vsn: sys/deprecated/netimp/raw_imp.c 4.14

+++ usr/src/sys/vax/vax/mem.c
@@ -1,4 +1,4 @@
-/*	mem.c	4.8	82/10/20	*/
+/*	mem.c	4.9	82/10/21	*/
 
 /*
  * Memory special file
@@ -64,7 +64,7 @@ mmrw(dev, uio, rw)
 			o = (int)uio->uio_offset & PGOFSET;
 			c = min((u_int)(NBPG - o), (u_int)iov->iov_len);
 			c = min(c, (u_int)(NBPG - ((int)iov->iov_base&PGOFSET)));
-			error = uiomove((caddr_t)&vmmap[o], c, rw, uio);
+			error = uiomove((caddr_t)&vmmap[o], (int)c, rw, uio);
 			continue;
 
 /* minor device 1 is kernel memory */
@@ -78,7 +78,7 @@ mmrw(dev, uio, rw)
 			c = iov->iov_len;
 			if (!kernacc((caddr_t)uio->uio_offset, c, rw == UIO_READ ? B_READ : B_WRITE))
 				goto fault;
-			error = uiomove((caddr_t)uio->uio_offset, c, rw, uio);
+			error = uiomove((caddr_t)uio->uio_offset, (int)c, rw, uio);
 			continue;
 
 /* minor device 2 is EOF/RATHOLE */
